# apps/api/tests/contracts/openapi.snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "IngestApproveRequest": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     }
    },
    "required": [
     "id"
    ],
    "title": "IngestApproveRequest",
    "type": "object"
   },
   "IngestRequestCreate": {
    "properties": {
     "content_type_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Content Type Id"
     },
     "contentid": {
      "maxLength": 32,
      "minLength": 1,
      "title": "Contentid",
      "type": "string"
     },
     "name": {
      "default": "",
      "maxLength": 200,
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "contentid"
    ],
    "title": "IngestRequestCreate",
    "type": "object"
   },
...
   "KeywordSearchItem": {
    "properties": {
     "addr1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Addr1"
     },
     "contentid": {
      "title": "Contentid",
      "type": "string"
     },
     "contenttypeid": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contenttypeid"
     },
     "firstimage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Firstimage"
     },
     "mapx": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mapx"
     },
     "mapy": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mapy"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "contentid",
     "title"
    ],
    "title": "KeywordSearchItem",
    "type": "object"
   },
   "KeywordSearchResponse": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/KeywordSearchItem"
      },
      "title": "Items",
      "type": "array"
     },
     "llm_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Status"
     },
     "rewritten": {
      "default": false,
      "title": "Rewritten",
      "type": "boolean"
     },
     "source": {
      "enum": [
       "tourapi",
       "unavailable"
      ],
      "title": "Source",
      "type": "string"
     }
    },
    "required": [
     "items",
     "source"
    ],
    "title": "KeywordSearchResponse",
    "type": "object"
   },
...
   "PlaceSearchItem": {
    "properties": {
     "address": {
      "title": "Address",
      "type": "string"
     },
     "category_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category Name"
     },
     "latitude": {
      "title": "Latitude",
      "type": "number"
     },
     "longitude": {
      "title": "Longitude",
      "type": "number"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "place_id": {
      "title": "Place Id",
      "type": "string"
     },
     "place_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Url"
     },
     "type": {
      "anyOf": [
       {
        "enum": [
         "cafe",
         "restaurant"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "required": [
     "place_id",
     "name",
     "latitude",
     "longitude",
     "address"
    ],
    "title": "PlaceSearchItem",
    "type": "object"
   },
   "PlaceSearchResponse": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PlaceSearchItem"
      },
      "title": "Items",
      "type": "array"
     },
     "source": {
      "enum": [
       "kakao",
       "unavailable"
      ],
      "title": "Source",
      "type": "string"
     }
    },
    "required": [
     "items",
     "source"
    ],
    "title": "PlaceSearchResponse",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "description": "NextSpot 관광 수요 분산·대안 장소 추천 AI 엔진 API",
  "title": "NextSpot API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/search/ingest-request": {
   "post": {
    "description": "\"다음 배치 추가 요청\" 큐잉 — admin_ingest_requests 에 pending 으로 upsert.\n\n무인증 + IP 당 분당 3회 제한. contentid 가 이미 요청돼 있으면(UNIQUE) 조용히 무시한다\n(ignore_duplicates=True — 상태와 무관하게 재요청이 기존 행을 덮어쓰지 않는다).",
    "operationId": "create_ingest_request_api_v1_search_ingest_request_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestRequestCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Ingest Request",
    "tags": [
     "search"
    ]
   }
  },
  "/api/v1/search/ingest-requests": {
   "get": {
    "operationId": "list_ingest_requests_api_v1_search_ingest_requests_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "default": "pending",
       "title": "Status",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Ingest Requests",
    "tags": [
     "search"
    ]
   }
  },
  "/api/v1/search/ingest-requests/approve": {
   "post": {
    "description": "적재 요청 승인 — detailCommon2/Intro2 조회 → facilities upsert → status='approved'.\n\nfacilities 적재가 실패하면 admin_ingest_requests 상태는 갱신하지 않는다(pending 유지 + 에러 반환).",
    "operationId": "approve_ingest_request_api_v1_search_ingest_requests_approve_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestApproveRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Approve Ingest Request",
    "tags": [
     "search"
    ]
   }
  },
  "/api/v1/search/keyword": {
   "get": {
    "description": "지도 검색 0건일 때만 프런트가 호출하는 TourAPI 키워드 폴백. 상위 5건만 반환한다.\n\n무인증 + IP 당 분당 12회 제한. TOURAPI_KEY 미설정/호출 실패는 500 이 아니라\n{items: [], source: 'unavailable'} 무해 폴백(events.py 축제 라우터와 동일 관례).\n경주(법정동 코드)로 검색하고, 경주 밖 장소는 주소·좌표로 한 번 더 거른다.\n정상 응답(source='tourapi')인데 0건이면 LLM 질의 재작성 폴백(P1-3)을 한 번 시도한다 —\n원 질의 우선 검색은 무개입, unavailable 위에는 LLM 을 쌓지 않는다.",
    "operationId": "search_keyword_endpoint_api_v1_search_keyword_get",
    "parameters": [
     {
      "description": "검색 키워드",
      "in": "query",
      "name": "q",
      "required": true,
      "schema": {
       "description": "검색 키워드",
       "maxLength": 100,
       "minLength": 1,
       "title": "Q",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/KeywordSearchResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Search Keyword Endpoint",
    "tags": [
     "search"
    ]
   }
  },
  "/api/v1/search/places": {
   "get": {
    "description": "장소·메뉴·업종으로 경주 카페/식당을 검색한다. 좌표는 같은 Kakao 레코드만 쓴다.",
    "operationId": "search_places_endpoint_api_v1_search_places_get",
    "parameters": [
     {
      "description": "장소·메뉴·업종 검색어",
      "in": "query",
      "name": "q",
      "required": true,
      "schema": {
       "description": "장소·메뉴·업종 검색어",
       "maxLength": 100,
       "minLength": 1,
       "title": "Q",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlaceSearchResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Search Places Endpoint",
    "tags": [
     "search"
    ]
   }
  },
...
 }
}